# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  /**
     * @minLength 3
     * @maxLength 20
     */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface UserProfile {
  id: number;
  username: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export type SaveDataSaveData = { [key: string]: unknown };

export interface SaveData {
  saveData: SaveDataSaveData;
  updatedAt: string;
}

export type SavePayloadSaveData = { [key: string]: unknown };

export interface SavePayload {
  saveData: SavePayloadSaveData;
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  tamerLevel: number;
  tamerName: string;
  collectionSize: number;
  updatedAt: string;
}

export interface PlayerProfile {
  username: string;
  tamerLevel: number;
  tamerName: string;
  collectionSize: number;
  topTeam: string[];
  updatedAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetLeaderboardParams = {
limit?: number;
};