# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: saves
    description: Cloud game saves
  - name: leaderboard
    description: Leaderboard
  - name: players
    description: Player profiles
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /saves:
    get:
      operationId: getSave
      tags: [saves]
      summary: Get cloud save for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Save data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveData"
        "404":
          description: No save found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: putSave
      tags: [saves]
      summary: Upload or update cloud save
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePayload"
      responses:
        "200":
          description: Save updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveData"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get top players ranked by tamer level
      parameters:
        - name: limit
          in: query
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  /players/{username}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get public profile of a player
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
        "404":
          description: Player not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthCredentials:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        password:
          type: string
          minLength: 6
      required:
        - username
        - password

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user

    UserProfile:
      type: object
      properties:
        id:
          type: number
        username:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - createdAt

    SaveData:
      type: object
      properties:
        saveData:
          type: object
          additionalProperties: true
        updatedAt:
          type: string
          format: date-time
      required:
        - saveData
        - updatedAt

    SavePayload:
      type: object
      properties:
        saveData:
          type: object
          additionalProperties: true
      required:
        - saveData

    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: number
        username:
          type: string
        tamerLevel:
          type: number
        tamerName:
          type: string
        collectionSize:
          type: number
        updatedAt:
          type: string
          format: date-time
      required:
        - rank
        - username
        - tamerLevel
        - tamerName
        - collectionSize
        - updatedAt

    PlayerProfile:
      type: object
      properties:
        username:
          type: string
        tamerLevel:
          type: number
        tamerName:
          type: string
        collectionSize:
          type: number
        topTeam:
          type: array
          items:
            type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - username
        - tamerLevel
        - tamerName
        - collectionSize
        - topTeam
        - updatedAt

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error